+++ src/components/square.hpp
@@ -11,8 +11,8 @@
  */
 struct Square
 {
-    int8_t file; // file from 0 ('a') to 7 ('h')
-    int8_t rank; // rank from 0 to 7
+    int file; // file from 0 ('a') to 7 ('h')
+    int rank; // rank from 0 to 7
 
     bool operator==(const Square &other) const {
         return file == other.file && rank == other.rank;
